Multiple fixes and improvements

+++ src/EventHandler/Media/AbstractVideo.php
@@ -11,7 +11,7 @@ use danog\MadelineProto\MTProto;
 abstract class AbstractVideo extends Media
 {
     /** Video duration in seconds */
-    public readonly int $duration;
+    public readonly float $duration;
     /** Whether the video supports streaming */
     public readonly bool $supportsStreaming;
     /** Video width */
